requirements_basic.txt
fastapi
uvicorn[standard]
python-multipart
whisper-ctranslate2
ffmpeg-python
aiofiles
websockets
